feat: convert relative paths to absolute (#108)
### Motivation

Most of LINZ's STAC items and collections are relative links as to make
it easier to move them around, we have a long term goal of converting
them all to absolute links but that is going to take a way.

In the mean time it would be great if this map tool supported relative
links and assets.

### Modification

I have added a symbol to all items that have been fetched via the
`fetchStac` to store their source location onto the STAC document, so
then any URL referenced from the document can converted with the source
URL.

I tried to avoid mutating the stac documents as I wasn't sure of the
imapct, so I avoided these two other ideas

- Converting the self link to be the actual source location then using
the self link to create relative
- Converting all links in the document to be absolute

---------

Co-authored-by: Pete Gadomski <pete.gadomski@gmail.com>